refactor(suite-native): refactor title into a separate component to optimize re-renders
