+++ suite-native/module-transactions/package.json
@@ -15,7 +15,6 @@
         "@react-navigation/native": "7.1.28",
         "@react-navigation/native-stack": "7.12.0",
         "@suite-common/dependency-injection": "workspace:*",
-        "@suite-common/discreet-mode": "workspace:*",
         "@suite-common/formatters": "workspace:*",
         "@suite-common/token-definitions": "workspace:*",
         "@suite-common/wallet-config": "workspace:*",
